refactor: split settings tabs by deployment mode
Reorganize settings to separate Cloud users from self-hosted users:

- Account tab → Engram Cloud info pane. "Coming soon" message + a
  "Switch to Self-hosted" shortcut for current users. When Cloud
  launches, sign-in lives here.
- Self-hosted tab → all current functionality:
  - Setup: link to github.com/Rasbandit/engram, server URL, test connection
  - Authentication: OAuth / API key flow (moved from Account)
  - Vault: vault picker (moved from Account)
  - About: version, source, license
  - Support development: Ko-fi button
- Advanced tab → adds Actions section at top (Sync now, Push all,
  Pull all). Universal across deployment modes.

TabContext gains switchToTab so any tab can route the user elsewhere.
Tab label corrected to sentence case ("Self-hosted").
