+++ src/settings.ts
@@ -69,15 +69,6 @@ export class EngramSyncSettingTab extends PluginSettingTab {
 		const tabBar = containerEl.createEl("nav", { cls: "engram-tab-bar" });
 		const contentEl = containerEl.createEl("section", { cls: "engram-tab-content" });
 
-		const ctx: TabContext = {
-			containerEl: contentEl,
-			app: this.app,
-			plugin: this.plugin,
-			redisplay: () => this.display(),
-			startDeviceFlow: () => this.startDeviceFlow(),
-			openProgressModal: () => this.openProgressModal(),
-		};
-
 		const activateTab = (tabId: string) => {
 			this.activeTab = tabId;
 			for (const btn of Array.from(tabBar.querySelectorAll<HTMLElement>(".engram-tab"))) {
@@ -88,6 +79,16 @@ export class EngramSyncSettingTab extends PluginSettingTab {
 			const btn = tabBar.querySelector<HTMLElement>(`[data-tab="${tab.id}"]`);
 			btn?.addClass("is-active");
 			tab.render({ ...ctx, containerEl: contentEl });
+		};
+
+		const ctx: TabContext = {
+			containerEl: contentEl,
+			app: this.app,
+			plugin: this.plugin,
+			redisplay: () => this.display(),
+			startDeviceFlow: () => this.startDeviceFlow(),
+			openProgressModal: () => this.openProgressModal(),
+			switchToTab: (id) => activateTab(id),
 		};
 
 		for (const tab of tabs) {
